Updated to Bytecoin Core 1.0.11. Added mandatory transaction extention

+++ extensions/multiply/files/CoinBaseConfiguration.cpp
@@ -77,8 +77,8 @@ void CoinBaseConfiguration::initOptions(boost::program_options::options_descript
     ("CRYPTONOTE_MINED_MONEY_UNLOCK_WINDOW", po::value<size_t>()->default_value(CryptoNote::parameters::CRYPTONOTE_MINED_MONEY_UNLOCK_WINDOW), "size_t")
     ("MAX_BLOCK_SIZE_INITIAL", po::value<uint64_t>()->default_value(CryptoNote::parameters::MAX_BLOCK_SIZE_INITIAL), "uint64_t")
     ("EXPECTED_NUMBER_OF_BLOCKS_PER_DAY", po::value<uint64_t>()->default_value(0), "uint64_t")
-    ("UPGRADE_HEIGHT_V2", po::value<uint64_t>()->default_value(0), "uint64_t")
-    ("UPGRADE_HEIGHT_V3", po::value<uint64_t>()->default_value(0), "uint64_t")
+    ("UPGRADE_HEIGHT_V2", po::value<uint32_t>()->default_value(0), "uint32_t")
+    ("UPGRADE_HEIGHT_V3", po::value<uint32_t>()->default_value(0), "uint32_t")
     ("DIFFICULTY_CUT", po::value<size_t>()->default_value(CryptoNote::parameters::DIFFICULTY_CUT), "size_t")
     ("DIFFICULTY_LAG", po::value<size_t>()->default_value(CryptoNote::parameters::DIFFICULTY_LAG), "size_t")
     ;
@@ -132,10 +132,10 @@ void CoinBaseConfiguration::init(const boost::program_options::variables_map& op
     EXPECTED_NUMBER_OF_BLOCKS_PER_DAY = options["EXPECTED_NUMBER_OF_BLOCKS_PER_DAY"].as<uint64_t>();
   }
   if (options.count("UPGRADE_HEIGHT_V2")) {
-    UPGRADE_HEIGHT_V2 = options["UPGRADE_HEIGHT_V2"].as<uint64_t>();
+    UPGRADE_HEIGHT_V2 = options["UPGRADE_HEIGHT_V2"].as<uint32_t>();
   }
   if (options.count("UPGRADE_HEIGHT_V3")) {
-    UPGRADE_HEIGHT_V3 = options["UPGRADE_HEIGHT_V3"].as<uint64_t>();
+    UPGRADE_HEIGHT_V3 = options["UPGRADE_HEIGHT_V3"].as<uint32_t>();
   }
   if (options.count("DIFFICULTY_CUT")) {
     DIFFICULTY_CUT = options["DIFFICULTY_CUT"].as<size_t>();
